#9 Refactor gradle configurations #7 Refactor project architecture / structure
