+++ blog/api/facade/build.gradle
@@ -1,0 +1,1 @@
+group 'com.excref.kotblog.blog.api'
