--- src/main/kotlin/com/excref/kotblog/domain/category/Category.kt
@@ -1,8 +1,0 @@
-package com.excref.kotblog.domain.category
-
-/**
- * @author Arthur Asatryan
- * @since 6/4/17 3:14 PM
- */
-class Category {
-}
