--- src/main/kotlin/com/excref/kotblog/domain/post/Post.kt
@@ -1,8 +1,0 @@
-package com.excref.kotblog.domain.post
-
-/**
- * @author Arthur Asatryan
- * @since 6/4/17 3:12 PM
- */
-class Post {
-}
