--- src/main/kotlin/com/excref/kotblog/domain/user/User.kt
@@ -1,8 +1,0 @@
-package com.excref.kotblog.domain.user
-
-/**
- * @author Arthur Asatryan
- * @since 6/4/17 3:11 PM
- */
-class User {
-}
