--- src/test/kotlin/com/excref/kotblog/KotblogApplicationTests.kt
@@ -1,20 +1,0 @@
-package com.excref.kotblog
-
-import org.junit.Test
-import org.junit.runner.RunWith
-import org.springframework.boot.test.context.SpringBootTest
-import org.springframework.test.context.junit4.SpringRunner
-
-/**
- * @author Arthur Asatryan
- * @since 5/4/17 1:17 PM
- */
-@RunWith(SpringRunner::class)
-@SpringBootTest
-class KotblogApplicationTests {
-
-    @Test
-    fun contextLoads() {
-    }
-
-}
