--- src/test/kotlin/com/excref/kotblog/service/tag/TagServiceIntegrationTest.kt
@@ -1,54 +1,0 @@
-package com.excref.kotblog.service.tag
-
-import com.excref.kotblog.service.test.AbstractServiceIntegrationTest
-import org.assertj.core.api.Assertions.assertThat
-import org.junit.Test
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.stereotype.Component
-
-/**
- * @author Arthur Asatryan
- * @since 6/4/17 4:12 PM
- */
-@Component
-class TagServiceIntegrationTest : AbstractServiceIntegrationTest() {
-
-    //region Dependencies
-    @Autowired
-    private lateinit var tagService: TagService
-    //endregion
-
-    //region Test methods
-    @Test
-    fun testCreate() {
-        // given
-        val name = "C++"
-        // when
-        val result = tagService.create(name)
-        // then
-        assertThat(result).isNotNull().extracting("name").containsOnly(name)
-    }
-
-    @Test
-    fun testGetByUuid() {
-        // given
-        val tag = helper.persistTag()
-        // when
-        val result = tagService.getByUuid(tag.uuid)
-        // then
-        assertThat(result).isNotNull().isEqualTo(tag)
-    }
-
-    @Test
-    fun testExistsForName() {
-        // given
-        val name = "C++"
-        helper.persistTag(name)
-        // when
-        val existsForName = tagService.existsForName(name)
-        // then
-        assertThat(existsForName).isTrue()
-    }
-    //endregion
-
-}
