--- src/test/kotlin/com/excref/kotblog/service/test/helper/ServiceImplTestHelper.kt
@@ -1,16 +1,0 @@
-package com.excref.kotblog.service.test.helper
-
-import com.excref.kotblog.domain.tag.Tag
-import java.util.*
-
-/**
- * @author Arthur Asatryan
- * @since 6/4/17 4:30 PM
- */
-class ServiceImplTestHelper {
-    //region Public methods
-    fun buildTag(): Tag = buildTag(UUID.randomUUID().toString())
-
-    fun buildTag(name: String): Tag = Tag(name)
-    //endregion
-}
